Move matchers into test-utils
