+++ test/@types/vitest.d.ts
@@ -9,11 +9,11 @@ import type { MoveResult } from "#enums/move-result";
 import type { EffectiveStat, PermanentStat } from "#enums/stat";
 import type { StatusEffect } from "#enums/status-effect";
 import type { WeatherType } from "#enums/weather-type";
-import type { ToHaveEffectiveStatMatcherOptions } from "#test/matchers/to-have-effective-stat-matcher";
-import type { ToHaveMoveResultMatcherOptions } from "#test/matchers/to-have-move-result-matcher";
-import type { ToHaveStatMatcherOptions } from "#test/matchers/to-have-stat-matcher";
-import type { ToHaveStatusEffectMatcherOptions } from "#test/matchers/to-have-status-effect-matcher";
-import type { ToHaveUsedMoveMatcherOptions } from "#test/matchers/to-have-used-move-matcher";
+import type { ToHaveEffectiveStatMatcherOptions } from "#test/test-utils/matchers/to-have-effective-stat-matcher";
+import type { ToHaveMoveResultMatcherOptions } from "#test/test-utils/matchers/to-have-move-result-matcher";
+import type { ToHaveStatMatcherOptions } from "#test/test-utils/matchers/to-have-stat-matcher";
+import type { ToHaveStatusEffectMatcherOptions } from "#test/test-utils/matchers/to-have-status-effect-matcher";
+import type { ToHaveUsedMoveMatcherOptions } from "#test/test-utils/matchers/to-have-used-move-matcher";
 import "vitest";
 
 declare module "vitest" {
